refactor: migrate sound_studio
